pyproject.toml
[project]
name = "DLAI"
version = "0.1.0"
description = "Learning LLM Models"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.8.4",
    "numpy>=1.26.4",
    "ollama>=0.6.1",
    "openai>=2.14.0",
    "pytest~=7.4.4",
    "pytest-html>=4.0.0",
    "pytest-html-reporter>=0.2.6",
    "python-dotenv>=1.2.1",
    "torch>=2.2.2",
]
